pyproject.toml
[tool.poetry]
name = "docx-hierarchy"
version = "0.1.0"
description = ""
authors = ["Madung2 <tuliphan91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-docx = "^1.1.2"
streamlit = "^1.35.0"
mammoth = "^1.8.0"
numbering = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
